add: reserve book functionality

+++ app/controllers/checkout_books_controller.rb
@@ -1,4 +1,6 @@
 class CheckoutBooksController < ApplicationController
+    before_action :required_login
+    before_action :book_exist?, only: [:create]
     before_action :already_borrowed?, only: [:create]
     before_action :book_is_available?, only: [:create]
     before_action :reach_maximum_limit?, only: [:create]
@@ -14,32 +16,55 @@ class CheckoutBooksController < ApplicationController
 
     def update
         book = Book.find(params[:id])
-        checkout_book = CheckoutBook.find_by(user_id: current_user.id, book_id: book.id)
-        if checkout_book.update(checkout_date: Date.current(), return_date: Date.current() + 5.days)
-            flash[:notice] = "Book successfully renewed"
-            redirect_to users_checkout_path
+        if book
+            checkout_book = CheckoutBook.find_by(user_id: current_user.id, book_id: book.id)
+            if checkout_book.update(checkout_date: Date.current(), return_date: Date.current() + 5.days)
+                flash[:notice] = "Book successfully renewed"
+                redirect_to users_checkout_path
+            else
+                flash[:alert] = "some error occured pleasetry again"
+                redirect_to users_checkout_pat
+            end 
         else
-            flash[:alert] = "some error occured pleasetry again"
-            redirect_to users_checkout_pat
+            flash[:alert] = "You can't perform this operation. details is wrong"
+            redirect_to request.referrer
         end 
     end
 
     def destroy
         book = Book.find(params[:id])
-        book.no_of_books_available = book.no_of_books_available + 1
-        book.save
-        checkout_book = CheckoutBook.find_by(user_id: current_user.id, book_id: book.id)
-        checkout_book.destroy
-        flash[:notice] = "Book successfully returned"
-        redirect_to users_checkout_path
+        if book
+            book.no_of_books_available = book.no_of_books_available + 1
+            book.save
+            checkout_book = CheckoutBook.find_by(user_id: current_user.id, book_id: book.id)
+            checkout_book.destroy
+            flash[:notice] = "Book successfully returned"
+            redirect_to users_checkout_path
+        else 
+            flash[:alert] = "You can't perform this operation. details is wrong"
+            redirect_to request.referrer
+        end
     end
 
     private 
+    def required_login
+        if !logged_in?
+            flash[:alert] = "You need to login before performing this action"
+            redirect_to login_path
+        end
+    end
 
     def already_borrowed? 
         @book = Book.find(params[:book])
         if current_user.books.include?(@book)
             flash[:alert] = "you already borrowed book!"
+            redirect_to request.referrer
+        end
+    end
+
+    def book_exist?
+        if !Book.find(params[:book])
+            flash[:alert] = "You can;t perform this operation. details is wrong"
             redirect_to request.referrer
         end
     end
